# Project/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/users/register":{"post":{"summary":"Register","operationId":"register_users_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/login":{"post":{"summary":"Login","operationId":"login_users_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_users_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/authenticate":{"post":{"summary":"Authenticate","operationId":"authenticate_users_authenticate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/shows":{"get":{"summary":"Get Shows","operationId":"get_shows_shows_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Add Shows","operationId":"add_shows_shows_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Show"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/shows/{show_id}":{"get":{"summary":"Get Id","operationId":"get_id_shows__show_id__get","parameters":[{"required":true,"schema":{"title":"Show Id","type":"string"},"name":"show_id","in":"path"},{"required":false,"schema":{"title":"Title","type":"string"},"name":"title","in":"query"},{"required":false,"schema":{"title":"Genre","type":"string"},"name":"genre","in":"query"},{"required":false,"schema":{"title":"Release Year","type":"integer"},"name":"release_year","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Show","operationId":"update_show_shows__show_id__put","parameters":[{"required":true,"schema":{"title":"Show Id","type":"string"},"name":"show_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Show"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Show","operationId":"delete_show_shows__show_id__delete","parameters":[{"required":true,"schema":{"title":"Show Id","type":"string"},"name":"show_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/episodes":{"get":{"summary":"Get Episode","operationId":"get_episode_episodes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/shows/{show_id}/episodes":{"get":{"summary":"Get List","operationId":"get_list_shows__show_id__episodes_get","parameters":[{"required":true,"schema":{"title":"Show Id","type":"string"},"name":"show_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Add Episodes","operationId":"add_episodes_shows__show_id__episodes_post","parameters":[{"required":true,"schema":{"title":"Show Id","type":"string"},"name":"show_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Episode"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/episodes/{episode_id}":{"put":{"summary":"Update Episode","operationId":"update_episode_episodes__episode_id__put","parameters":[{"required":true,"schema":{"title":"Episode Id","type":"string"},"name":"episode_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Episode"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Episode","operationId":"delete_episode_episodes__episode_id__delete","parameters":[{"required":true,"schema":{"title":"Episode Id","type":"string"},"name":"episode_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/me/watchlist":{"get":{"summary":"Get Watchlist","operationId":"get_watchlist_me_watchlist_get","parameters":[{"required":false,"schema":{"title":"Status","type":"string"},"name":"status","in":"query"},{"required":false,"schema":{"title":"Min Rating","type":"integer"},"name":"min_rating","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"summary":"Add Watchlist","operationId":"add_watchlist_me_watchlist_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Watchlist"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/me/watchlist/{watchlist_id}":{"delete":{"summary":"Delete Watchlist","operationId":"delete_watchlist_me_watchlist__watchlist_id__delete","parameters":[{"required":true,"schema":{"title":"Watchlist Id","type":"string"},"name":"watchlist_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"patch":{"summary":"Update Watchlist","operationId":"update_watchlist_me_watchlist__watchlist_id__patch","parameters":[{"required":true,"schema":{"title":"Watchlist Id","type":"string"},"name":"watchlist_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Watchlist"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/me/watchlist/{watchlist_id}/watched-episodes":{"get":{"summary":"Get Watched Episodes","operationId":"get_watched_episodes_me_watchlist__watchlist_id__watched_episodes_get","parameters":[{"required":true,"schema":{"title":"Watchlist Id","type":"string"},"name":"watchlist_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/me/watchlist/{watchlist_id}/watched-episodes/{episode_id}":{"post":{"summary":"Add Watched","operationId":"add_watched_me_watchlist__watchlist_id__watched_episodes__episode_id__post","parameters":[{"required":true,"schema":{"title":"Watchlist Id","type":"string"},"name":"watchlist_id","in":"path"},{"required":true,"schema":{"title":"Episode Id","type":"string"},"name":"episode_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchedEpisode"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"summary":"Delete Watchlist Episode","operationId":"delete_watchlist_episode_me_watchlist__watchlist_id__watched_episodes__episode_id__delete","parameters":[{"required":true,"schema":{"title":"Watchlist Id","type":"string"},"name":"watchlist_id","in":"path"},{"required":true,"schema":{"title":"Episode Id","type":"string"},"name":"episode_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_login_users_login_post":{"title":"Body_login_users_login_post","required":["username","password"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"}}},"Episode":{"title":"Episode","required":["show_id","season_number","episode_number","title","duration_minutes"],"type":"object","properties":{"show_id":{"title":"Show Id","type":"string"},"season_number":{"title":"Season Number","type":"integer"},"episode_number":{"title":"Episode Number","type":"integer"},"title":{"title":"Title","type":"string"},"duration_minutes":{"title":"Duration Minutes","type":"integer"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Show":{"title":"Show","required":["title","description","genre","release_year","type"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"description":{"title":"Description","type":"string"},"genre":{"title":"Genre","type":"string"},"release_year":{"title":"Release Year","type":"integer"},"type":{"title":"Type","type":"string"}}},"User":{"title":"User","required":["username","email","password"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"email":{"title":"Email","type":"string"},"password":{"title":"Password","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}},"WatchedEpisode":{"title":"WatchedEpisode","required":["episode_id","watched_at"],"type":"object","properties":{"episode_id":{"title":"Episode Id","type":"string"},"watched_at":{"title":"Watched At","type":"integer"}}},"Watchlist":{"title":"Watchlist","required":["show_id","status","rating","notes"],"type":"object","properties":{"show_id":{"title":"Show Id","type":"string"},"status":{"title":"Status","type":"string"},"rating":{"title":"Rating","type":"integer"},"notes":{"title":"Notes","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/users/login"}}}}}}